adds freeze and fixes small issues

+++ src/main/java/com/softropic/sendam/gateway/sms/service/SmsProviderReportListener.java
@@ -120,7 +120,7 @@ public class SmsProviderReportListener {
             parent.getSendRequestId(), 
             summaries,
             parent.getReservationId(),
-            (long) totalActualSegments
+            totalActualSegments
         ));
     }
